fix: address all remaining review issues (important + medium severity)
Important fixes:
- Issue 9: Dispose IMAP client on failed connect/auth retries
- Issue 10: Add 5s timeout to MachineId.RunProcess with kill fallback
- Issue 11: Use InvariantCulture + RoundtripKind for DateTime.Parse in QueueWorker
- Issue 12: Replace .First() with .FirstOrDefault() + descriptive exception in all executors
- Issue 13: Validate account ID exists in QueueManager.EnqueueSend
- Issue 14: Atomic conditional UPDATE in QueueRepository.Cancel (TOCTOU fix)
- Issue 15: SmtpConnectionManager: add IsAuthenticated check, retry, logging
- Issue 16: Change UID from int to long throughout MessageRepository/FolderRepository
- Issue 17: DashboardHub SendToAll now async void with try/catch
- Issue 18: Log warning when config file not found, falling back to defaults
- Issue 19: Throw on null password instead of falling back to empty string

Medium fixes:
- M1: OrganizeTools.ParseUids uses TryParse, returns error JSON on invalid input
- M2: DashboardAuthRepository.UpsertPin uses atomic DELETE+INSERT
- M3: AcpEmailAnalyzer.EnsureSessionAsync guarded by SemaphoreSlim
- M4: AccountsApi test endpoint returns 502 on connection failure
- M5: ConfigLoader throws on null deserialization instead of silent fallback
- M6: SyncManager sync failures logged as Error instead of Warning
- M7: ImapSyncService snippet catch narrowed to specific exception types
- M8: SyncApi adds general catch returning 500
- M10: CacheEvictor Task.Delay moved inside try-catch

+++ src/UltimateImapMcp.ImapClient/Repositories/FolderRepository.cs
@@ -5,7 +5,7 @@ namespace UltimateImapMcp.ImapClient.Repositories;
 public record FolderRecord(
     int Id, string AccountId, string Path, string? DisplayName,
     string? Role, string Delimiter, string? Flags,
-    int MessageCount, int UnreadCount, int LastSyncedUid,
+    int MessageCount, int UnreadCount, long LastSyncedUid,
     string? LastSyncedAt, bool SyncEnabled, bool IdleEnabled, int PollInterval);
 
 public class FolderRepository(AppDatabase db)
@@ -53,7 +53,7 @@ public class FolderRepository(AppDatabase db)
         return list;
     }
 
-    public void UpdateSyncState(int folderId, int lastSyncedUid, int messageCount, int unreadCount)
+    public void UpdateSyncState(int folderId, long lastSyncedUid, int messageCount, int unreadCount)
     {
         db.ExecuteWrite(conn =>
         {
@@ -81,7 +81,7 @@ public class FolderRepository(AppDatabase db)
         Flags: r.IsDBNull(r.GetOrdinal("flags")) ? null : r.GetString(r.GetOrdinal("flags")),
         MessageCount: r.GetInt32(r.GetOrdinal("message_count")),
         UnreadCount: r.GetInt32(r.GetOrdinal("unread_count")),
-        LastSyncedUid: r.GetInt32(r.GetOrdinal("last_synced_uid")),
+        LastSyncedUid: r.GetInt64(r.GetOrdinal("last_synced_uid")),
         LastSyncedAt: r.IsDBNull(r.GetOrdinal("last_synced_at")) ? null : r.GetString(r.GetOrdinal("last_synced_at")),
         SyncEnabled: r.GetInt32(r.GetOrdinal("sync_enabled")) == 1,
         IdleEnabled: r.GetInt32(r.GetOrdinal("idle_enabled")) == 1,
